+++ NEWS.md
@@ -1,5 +1,9 @@
 # worldfootballR 0.5.12.4000
 
+### New Functions
+
+* `fb_squad_wages()` designed to get player wage estimates from FBref via Capology
+
 ### Bugs
 
 * `load_fb_big5_advanced_season_stats()` returning zero row df when no season selected. Now rectified